backend/requirements.txt
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djoser==2.0.5
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
webcolors==1.12
reportlab==3.6.12